Fix clippy and increase timeout

+++ dds/src/implementation/rtps_udp_psm/mapping_rtps_messages/submessages/data.rs
@@ -105,7 +105,7 @@ impl<'de: 'a, 'a> MappingReadSubmessage<'de> for DataSubmessage<'a> {
                 - octets_to_inline_qos as usize
                 - 4
                 - inline_qos_len;
-            let (data, following) = buf.split_at(serialized_payload_length as usize);
+            let (data, following) = buf.split_at(serialized_payload_length);
             *buf = following;
             SerializedPayload::new(data)
         } else {
